% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bttdSoftImpute.R
\name{bttdSoftImpute}
\alias{bttdSoftImpute}
\title{BTTSoftImpute}
\usage{
bttdSoftImpute(Y, lambda = 0, rank_Y = 2, ttrank_max = 100,
  tol_dx = 1e-05, tol_df = 1e-05, tol_f = 0, maxit = 100,
  verbose = FALSE, X0 = NULL, method = "ALS")
}
\arguments{
\item{Y}{A sparse matrix format with attributes
$i, $j, $v, $nrow, $ncol, and $dimnames,
which are [nnz]-vector, [nnz x N]-matrix, [nnz]-vector, scalar,
[N]-vector, and list, respectively.}

\item{rank_Y}{The rank of the estimated low rank matrix, X.}

\item{ttrank_max}{The maximal TT rank of right singular vectors, X$v.}
}
\value{
X = list(u, d, v), v is in block TT format
}
\description{
BTTSoftImpute implements block tensor train decomposition for
missing data estimation. It extends the SoftImpute method
for matrix data imputation to higher-order tensor data.
}
\details{
Last modified: 2018.01.15. by Namgil Lee (Kangwon National University)
}
