% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bttSparseSVD_SVD.R
\name{bttSparseSVD_SVD}
\alias{bttSparseSVD_SVD}
\title{bttSparseSVD_SVD}
\usage{
bttSparseSVD_SVD(Afill, ttrank_max, tol_u, lambda, maxit, verbose,
  X0 = NULL, final.svd = FALSE)
}
\arguments{
\item{Afill}{list(Ares, X), i.e., SPLR format}

\item{lambda}{parameter for soft-thresholding}

\item{Ares}{SP format, with elements $i, $j, $v, $nrow, $ncol}

\item{X}{LR format, with elements $u, $d, $v. X$v is in BlockTT format.}
}
\value{
X: LR format
}
\description{
Low rank tensor approximation (SVD) based on block TT format.
}
\details{
We use SVDS to optimize each core tensors simulaneously with
left-singular-vectors(U), i.e.,
(U,V1) -> (U,V2) -> ...

Last modified: 2017.12.31
}
