% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_sparse_matrix.R
\name{reduce_to_sparse_matrix}
\alias{reduce_to_sparse_matrix}
\title{reduce_to_sparse_matrix}
\usage{
reduce_to_sparse_matrix(X)
}
\arguments{
\item{X}{X has attributes $i, $j, $v, $nrow, $ncol, $dimnames}
}
\value{
Output X has the same attributes as the input X, But $i and $j are forced to be in vector instead of matrix, and $nrow and $ncol are forced to be scalars instead of vectors.
}
\description{
Convert a sparse tensor X into the sparse matrix format.
}
